arxiv_assistant/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10  # Required for ChatGroq and other integrations
langchain-groq>=0.0.1  # Groq integration
langgraph>=0.0.10
chromadb>=0.4.18
# NOTE: Install PyTorch manually following instructions at https://pytorch.org/get-started/locally/
numpy==1.24.3  # Must use 1.x series for PyTorch compatibility
huggingface_hub==0.16.4  # Needed for sentence-transformers compatibility
sentence-transformers==2.2.2
arxiv>=1.4.8
pypdf>=3.17.0
python-dotenv>=1.0.0
groq>=0.4.0
httpx>=0.25.0
ijson>=3.2.0  # For streaming JSON parsing
tqdm>=4.64.0  # For progress bars 